Geometry1D created
