+++ autoarray/mask/mask_1d.py
@@ -10,6 +10,7 @@ if TYPE_CHECKING:
 from autoarray.mask.abstract_mask import Mask
 
 from autoarray import exc
+from autoarray.geometry.geometry_1d import Geometry1D
 from autoarray.structures.arrays import array_1d_util
 from autoarray.structures.grids import grid_1d_util
 from autoarray import type as ty
@@ -66,6 +67,12 @@ class Mask1D(Mask):
             pass
         else:
             self.origin = (0.0,)
+
+    @property
+    def geometry(self):
+        return Geometry1D(
+            shape_native=self.shape_native, pixel_scales=self.pixel_scales, origin=self.origin
+        )
 
     @classmethod
     def manual(
@@ -217,22 +224,6 @@ class Mask1D(Mask):
     @property
     def shape_slim(self) -> Tuple[int]:
         return self.shape
-
-    @property
-    def shape_slim_scaled(self) -> Tuple[float]:
-        return (float(self.pixel_scales[0] * self.shape_slim[0]),)
-
-    @property
-    def scaled_maxima(self) -> Tuple[float]:
-        return (float(self.shape_slim_scaled[0] / 2.0 + self.origin[0]),)
-
-    @property
-    def scaled_minima(self) -> Tuple[float]:
-        return (-float(self.shape_slim_scaled[0] / 2.0) + self.origin[0],)
-
-    @property
-    def extent(self):
-        return np.array([self.scaled_minima[0], self.scaled_maxima[0]])
 
     def output_to_fits(self, file_path: str, overwrite: bool = False):
         """
